1556: Extract storeColorPicker and general customColorPicker module
