+++ visualization/app/codeCharta/ui/colorSettingsPanel/colorSettingsPanel.component.html
@@ -1,23 +1,23 @@
 <range-slider-component></range-slider-component>
 
 <div ng-if="!$ctrl._viewModel.isDeltaState">
-	<cc-map-color-picker map-color-for="positive" label="0 to &lt; {{ $ctrl._viewModel.colorRange.from | number: 0 }}">
-	</cc-map-color-picker>
-	<cc-map-color-picker
+	<cc-store-color-picker map-color-for="positive" label="0 to &lt; {{ $ctrl._viewModel.colorRange.from | number: 0 }}">
+	</cc-store-color-picker>
+	<cc-store-color-picker
 		map-color-for="neutral"
 		label="&ge; {{ $ctrl._viewModel.colorRange.from | number: 0 }} to &le; {{ $ctrl._viewModel.colorRange.to | number: 0 }}"
 	>
-	</cc-map-color-picker>
-	<cc-map-color-picker map-color-for="negative" label="&gt; {{ $ctrl._viewModel.colorRange.to | number: 0 }} to Infinite">
-	</cc-map-color-picker>
+	</cc-store-color-picker>
+	<cc-store-color-picker map-color-for="negative" label="&gt; {{ $ctrl._viewModel.colorRange.to | number: 0 }} to Infinite">
+	</cc-store-color-picker>
 </div>
 
 <div ng-if="$ctrl._viewModel.isDeltaState">
-	<cc-map-color-picker map-color-for="positiveDelta" label="+&Delta;"> </cc-map-color-picker>
-	<cc-map-color-picker map-color-for="negativeDelta" label="-&Delta;"> </cc-map-color-picker>
+	<cc-store-color-picker map-color-for="positiveDelta" label="+&Delta;"> </cc-store-color-picker>
+	<cc-store-color-picker map-color-for="negativeDelta" label="-&Delta;"> </cc-store-color-picker>
 </div>
 
-<cc-map-color-picker map-color-for="selected" label="selected"> </cc-map-color-picker>
+<cc-store-color-picker map-color-for="selected" label="selected"> </cc-store-color-picker>
 
 <md-input-container class="md-block" ng-if="!$ctrl._viewModel.isDeltaState">
 	<md-checkbox class="md-primary" ng-model="$ctrl._viewModel.invertColorRange" ng-change="$ctrl.invertColorRange()">
