+++ visualization/app/codeCharta/ui/customColorPicker/customColorPicker.scss
@@ -1,0 +1,31 @@
+@import "~angularjs-color-picker/dist/angularjs-color-picker.min.css";
+
+color-picker {
+	display: inline-block;
+	vertical-align: middle;
+
+	.cc-color-picker-swatch-brush-icon {
+		position: absolute;
+		padding: 2px;
+	}
+
+	.color-picker-wrapper .color-picker-input-swatch {
+		margin-top: 8px;
+		width: 100%;
+	}
+
+	.color-picker-panel {
+		bottom: -92px; /* centered vertically to swatch */
+	}
+
+	.color-picker-wrapper .color-picker-panel .color-picker-grid-wrapper .color-picker-row .color-picker-slider {
+		width: 10px;
+		height: 4px;
+		margin-top: -3px;
+	}
+
+	.cc-color-picker-grid-margin {
+		display: table-cell;
+		width: 16px;
+	}
+}
