+++ visualization/app/codeCharta/ui/customColorPicker/storeColorPicker/colorHelper.spec.ts
