+++ visualization/app/codeCharta/ui/customColorPicker/storeColorPicker/colorHelper.ts
