+++ visualization/app/codeCharta/ui/customColorPicker/storeColorPicker/storeColorPicker.component.html
@@ -1,4 +1,4 @@
-<div class="cc-map-color-picker-wrapper" ng-click="colorPickerApi.open()">
+<div class="cc-store-color-picker-wrapper" ng-click="colorPickerApi.open()">
 	<color-picker ng-model="color" api="colorPickerApi" options="colorPickerOptions" event-api="colorPickerEventApi"> </color-picker>
 	<p class="cc-color-picker-label">{{ $ctrl.label }}</p>
 </div>
