+++ visualization/app/codeCharta/ui/edgeSettingsPanel/edgeSettingsPanel.component.html
@@ -61,8 +61,8 @@
 	</md-input-container>
 </md-slider-container>
 
-<cc-map-color-picker map-color-for="outgoingEdge" label="Outgoing Edge"> </cc-map-color-picker>
-<cc-map-color-picker map-color-for="incomingEdge" label="Incoming Edge"> </cc-map-color-picker>
+<cc-store-color-picker map-color-for="outgoingEdge" label="Outgoing Edge"> </cc-store-color-picker>
+<cc-store-color-picker map-color-for="incomingEdge" label="Incoming Edge"> </cc-store-color-picker>
 
 <md-input-container class="md-block">
 	<md-checkbox
