+++ visualization/app/codeCharta/ui/legendPanel/legendPanel.component.scss
@@ -81,7 +81,7 @@ legend-panel-component {
 			font-weight: 600;
 		}
 
-		cc-map-color-picker .color-picker-panel {
+		cc-store-color-picker .color-picker-panel {
 			/* prevent panel to be below browser bottom */
 			bottom: -50px;
 		}
