--- visualization/app/codeCharta/ui/mapColorPicker/mapColorPicker.component.scss
@@ -1,62 +1,0 @@
-@import "~angularjs-color-picker/dist/angularjs-color-picker.min.css";
-
-cc-map-color-picker {
-	.cc-map-color-picker-wrapper {
-		margin: 4px 0;
-
-		&:hover {
-			cursor: pointer;
-		}
-
-		.color-picker-wrapper .color-picker-swatch {
-			width: 16px;
-			height: 16px;
-			padding: 0;
-		}
-
-		.color-picker-wrapper .color-picker-input-swatch {
-			margin-top: 8px;
-			width: 100%;
-		}
-
-		.cc-color-picker-swatch-brush-icon {
-			position: absolute;
-			width: 16px;
-			height: 16px;
-			padding: 2px;
-			opacity: 0;
-		}
-
-		&:hover .cc-color-picker-swatch-brush-icon,
-		.color-picker-open .cc-color-picker-swatch-brush-icon {
-			opacity: 1;
-		}
-	}
-
-	.cc-color-picker-label {
-		display: inline-block;
-		margin: 0;
-		vertical-align: middle;
-	}
-
-	color-picker {
-		display: inline-block;
-		vertical-align: middle;
-
-		.color-picker-panel {
-			right: 24px; /* left to swatch */
-			bottom: -92px; /* centered vertically to swatch */
-		}
-
-		.color-picker-wrapper .color-picker-panel .color-picker-grid-wrapper .color-picker-row .color-picker-slider {
-			width: 10px;
-			height: 4px;
-			margin-top: -3px;
-		}
-
-		.cc-color-picker-grid-margin {
-			display: table-cell;
-			width: 16px;
-		}
-	}
-}
