+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.component.scss
@@ -185,7 +185,7 @@ ribbon-bar-component {
 		}
 	}
 
-	cc-map-color-picker .cc-map-color-picker-wrapper {
+	cc-store-color-picker .cc-store-color-picker-wrapper {
 		margin-left: 5px;
 	}
 }
