+++ visualization/app/codeCharta/ui/ui.ts
@@ -3,7 +3,7 @@
 import angular from "angular"
 
 // Plop: Append module import here
-import "./mapColorPicker/mapColorPicker.module"
+import "./customColorPicker/customColorPicker.module"
 import "./customConfigs/customConfigs.module"
 import "./attributeTypeSelector/attributeTypeSelector.module"
 import "./sortingOption/sortingOption.module"
@@ -49,7 +49,7 @@ import "./maxTreeMapFiles/maxTreeMapFiles.module"
 
 angular.module("app.codeCharta.ui", [
 	// Plop: Append component name here
-	"app.codeCharta.ui.mapColorPicker",
+	"app.codeCharta.ui.customColorPicker",
 	"app.codeCharta.ui.customConfigs",
 	"app.codeCharta.ui.sortingOptionDialog",
 	"app.codeCharta.ui.sortingButton",
